pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skills-mcp-server"
version = "0.1.0.dev0"
description = "An open-source MCP server that hosts SKILL.md skill bundles and serves them to any MCP-capable agent."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "Darshan Yadav" },
]
keywords = ["mcp", "skills", "llm", "agent", "claude"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "mcp",
    "PyYAML",
    "pydantic>=2",
    "GitPython",
    "aiohttp",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[project.scripts]
skills-mcp-server = "skills_mcp_server.__main__:main"

[project.urls]
Homepage = "https://github.com/darshanyadav/skills-mcp-server"
Repository = "https://github.com/darshanyadav/skills-mcp-server"
Issues = "https://github.com/darshanyadav/skills-mcp-server/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/skills_mcp_server"]

[tool.hatch.build.targets.sdist]
include = [
    "src/skills_mcp_server",
    "README.md",
    "SPEC.md",
    "pyproject.toml",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
